+++ tools/poke.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+TERM=linux /usr/bin/setterm --blank=poke < /dev/console