+++ tools/setblank.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+TERM=linux /usr/bin/setterm -blank 5 < /dev/tty1 > /dev/tty1